feat: add component Promo
