+++ flower_shop/src/components/Promo/promo.scss
