+++ flower_shop/src/components/TotalPrice/TotalPrice.scss
@@ -1,12 +1,16 @@
 .total_content {
     width: 270px;
-    height: 385px;
+    height: 360px;
 
     border-radius: 6px;
     background: $base-bg-color;
     box-shadow: 0px 2px 30px 0px rgba(0, 0, 0, 0.08);
 
     padding: 25px;
+
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
 }
 
 .total_title {
@@ -44,7 +48,11 @@
     font-weight: 700;
     line-height: 16px;
     position: relative;
-    margin-top: 30px;
+    margin-top: 50px;
+    margin-bottom: 40px;
+
+    display: flex;
+    justify-content: space-between;
 }
 
 .invoice_total::after {
@@ -54,5 +62,5 @@
     background-color: #EFEFEF;
     display: block;
     position: absolute;
-    top: 40px;
+    top: -30px;
 }
